requirements.txt
qrcode
qrtools==0.0.2
zbar-py
pillow
zbarlight
pyzbar
